clang tidy warnings

+++ src/controllers/commands/builtin/chatterino/Debugging.cpp
@@ -145,8 +145,9 @@ QString listEnvironmentVariables(const CommandContext &ctx)
         "recentMessagesApiUrl: " + env.recentMessagesApiUrl,
         "linkResolverUrl: " + env.linkResolverUrl,
         "proxyUrl: " + env.proxyUrl.value_or("N/A"),
-        "proxyTwitch: " + QString::number(env.proxyTwitch),
-        "proxyTwitchApiOnly: " + QString::number(env.proxyTwitchApiOnly),
+        "proxyTwitch: " + QString::number(static_cast<int>(env.proxyTwitch)),
+        "proxyTwitchApiOnly: " +
+            QString::number(static_cast<int>(env.proxyTwitchApiOnly)),
         "twitchServerHost: " + env.twitchServerHost,
         "twitchServerPort: " + QString::number(env.twitchServerPort),
         "twitchServerSecure: " + QString::number(env.twitchServerSecure),
